+++ src/controller/CameraController.h
@@ -20,8 +20,8 @@ public:
         Node& node) noexcept override;
 
     virtual void onKey(Input* input, const ki::RenderClock& clock) override;
-    virtual void onMouseMove(Input* input, double xoffset, double yoffset) override;
-    virtual void onMouseScroll(Input* input, double xoffset, double yoffset) override;
+    virtual void onMouseMove(Input* input, float xoffset, float yoffset) override;
+    virtual void onMouseScroll(Input* input, float xoffset, float yoffset) override;
 
 private:
     Node* m_node{ nullptr };
