+++ src/gui/Input.h
@@ -42,15 +42,15 @@ public:
 
     bool isModifierPressed(Modifier modifier);
 
-    void onMouseMove(double xpos, double ypos);
+    void onMouseMove(float xpos, float ypos);
     void onMouseButton(int button, int action, int modifiers);
 
 public:
-    double mouseX = 0;
-    double mouseY = 0;
+    float mouseX = 0;
+    float mouseY = 0;
 
-    double mouseXoffset = 0;
-    double mouseYoffset = 0;
+    float mouseXoffset = 0;
+    float mouseYoffset = 0;
 
     Window* window;
 
@@ -68,4 +68,3 @@ private:
     bool m_firstMouse = true;
 
 };
-
