+++ src/model/Viewport.cpp
@@ -55,7 +55,7 @@ Viewport::Viewport(
 
     setUpdate([this, origPosition, origRotation](Viewport& vp, const UpdateContext& ctx) {
         glm::vec3 rotation{ origRotation};
-        rotation.y = 5.f * sin(ctx.m_clock.ts);
+        rotation.y = 5.f * sinf(static_cast<float>(ctx.m_clock.ts));
 
         glm::vec3 position{ origPosition };
         position.z += 0.001f;
